make GIG vignette considerably faster (via save file *.rds)
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/copula@1839 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/copula/Misc/timing_vignettes.R
@@ -1,4 +1,4 @@
-## By Marius Hofert
+## Idea, etc: By Marius Hofert
 
 ## Timing vignettes
 
@@ -8,37 +8,45 @@ setwd("../vignettes")
 ## Timing command
 timer <- function(x) {
     tm <- system.time(res <- tryCatch(x, error = function(e) e))
-    if(is(res, "simpleError")) NA else round(tm[["elapsed"]], 2)
+    if(inherits(res, "error")) NA else round(tm[["elapsed"]], 2)
 }
+do.rm <- !(Sys.info()[["user"]] %in% c("maechler", "<you>"))#
 
 ## Determine all vignettes
 all <- system("ls", intern = TRUE)
-all.Rnw <- all[grepl(".Rnw", all)]
-all.Rmd <- all[grepl(".Rmd", all)]
-all.vign <- c(all.Rnw, all.Rmd)
-n <- length(all.vign)
+(all.Rnw <- all[grepl("\\.Rnw$", all)])
+(all.Rmd <- all[grepl("\\.Rmd$", all)])
+(n <- length(all.vign <- c(all.Rnw, all.Rmd)))
 if(n < 1)
     stop("Found no vignettes in the current working directory")
 
 ## Time measurement
-## Careful with environments: the code is executed in parent.frame(); see, e.g., ?render.
-## For example, if a global counter 'i' is used here, it is overwritten by wild_animals.Rmd.
-## If new.env() is used in render(), then dNAC.Rmd and AR_Clayton.Rmd fail with weird problems.
-tm <- c()
+tm <- numeric(n)
 for(j in seq_along(all.Rnw)) { # iterate over all .Rnw
-    tm <- c(tm, timer(system(paste("R CMD Sweave --pdf", all.Rnw[j]))))
-    Rnwbasename <- tools::file_path_sans_ext(basename(all.Rnw[j]))
+    cat("==> File ", fil <- all.Rnw[j],": ")
+    tm[j] <- timer(system(paste("R CMD Sweave --pdf", fil)))
+    Rnwbasename <- tools::file_path_sans_ext(basename(fil))
     all <- system(paste0("ls ",Rnwbasename,"*"), intern = TRUE)
-    all <- all[all != all.Rnw[j]] # keep original file
-    file.remove(all)
+    if(do.rm) file.remove(all[!grepl(fil, all, fixed=TRUE)])# keep *.Rnw file (and backups of that)
+    cat("\n \\--> finished ", fil, "\n-==-==-==-==-\n")
 }
+n1 <- length(all.Rnw)
+## rmarkdown::render() :
+## - Careful with environments: the code is executed in parent.frame(); see, ?render.
+## - For example, if a global counter 'i' is used here, it is overwritten by wild_animals.Rmd.
+## - If new.env() is used in render(), then dNAC.Rmd and AR_Clayton.Rmd fail with weird problems.
 for(j in seq_along(all.Rmd)) { # iterate over all .Rmd
-    tm <- c(tm, timer(rmarkdown::render(all.Rmd[j])))
-    Rmdbasename <- tools::file_path_sans_ext(basename(all.Rmd[j]))
-    file.remove(paste0(Rmdbasename,".html"))
+    cat("==> File ", fil <- all.Rmd[j],": ")
+    tm[j+n1] <- timer(rmarkdown::render(fil))
+    Rmdbasename <- tools::file_path_sans_ext(basename(fil))
+    if(do.rm) file.remove(paste0(Rmdbasename,".html"))
+    cat("\n-==-==-==-==-\n finished ", fil, "\n")
 }
-file.remove("Rplots.pdf")
+if(do.rm) file.remove("Rplots.pdf")
 
 ## Results
-timings <- data.frame("Vignette" = all.vign, "Elapsed.time.in.sec" = tm)
+timings <- data.frame("Vignette" = all.vign,
+                      "Elapsed.time.in.sec" = tm)
+sink("../Misc/timings.out", split = TRUE)
 timings[order(timings[,"Elapsed.time.in.sec"], decreasing = TRUE), ]
+sink()
